# [SPARK-45606][SQL] Release restrictions on multi-layer runtime filter