### What changes were proposed in this pull request?
Before https://github.com/apache/spark/pull/39170, Spark only supports insert runtime filter for application side of shuffle join on single-layer. Considered it's not worth to insert more runtime filter if the column already exists runtime filter, Spark restricts it at https://github.com/apache/spark/blob/7057952f6bc2c5cf97dd408effd1b18bee1cb8f4/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/InjectRuntimeFilter.scala#L346
For example
`select * from bf1 join bf2 on bf1.c1 = bf2.c2 and bf1.c1 = bf2.b2 where bf2.a2 = 62`
This SQL have two join conditions. There will insert two runtime filter on `bf1.c1` if haven't the restriction mentioned above.
At that time, it was reasonable.

After https://github.com/apache/spark/pull/39170, Spark supports insert runtime filter for one side of any shuffle join on multi-layer. But the restrictions on multi-layer runtime filter mentioned above looks outdated.
For example
`select * from bf1 join bf2 join bf3 on bf1.c1 = bf2.c2 and bf3.c3 = bf1.c1 where bf2.a2 = 5`
Assume bf2 as the build side and insert a runtime filter for bf1. We can't insert the same runtime filter for bf3 due to there are already a runtime filter on `bf1.c1`.
The behavior is different from the origin and is unexpected.

The change of the PR doesn't affect the restriction mentioned above.


### Why are the changes needed?
Release restrictions on multi-layer runtime filter.
Expand optimization surface.


### Does this PR introduce _any_ user-facing change?
'No'.
New feature.


### How was this patch tested?
Test cases updated.

Micro benchmark for q9 in TPC-H.
**TPC-H 100**
Query | Master(ms) | PR(ms) | Difference(ms) | Percent
-- | -- | -- | -- | --
q9 | 26491 | 20725 | 5766| 27.82%


### Was this patch authored or co-authored using generative AI tooling?
'No'.